Se agrego funcionalidad para el flotante

+++ js/index.js
@@ -14,7 +14,7 @@ document.addEventListener("DOMContentLoaded", function () {
         form.querySelector(".buy-button").addEventListener("click", function (event) {
             event.preventDefault();
 
-            const quantity = parseInt(form.querySelector("input[name='quantity']").value);
+            const quantity = parseFloat(form.querySelector("input[name='quantity']").value);
             const product_name = form.querySelector("input[name='product_name']").value;
             const product_price = parseFloat(form.querySelector("input[name='product_price']").value);
             const total_price = quantity * product_price;
